fix(mcp): add optional JSON Schema validation for MCP tool parameters
Extract validation logic from tools.rs into a dedicated validation
module. When the `mcp-validation` feature is enabled, tool argument
validation delegates to the `jsonschema` crate for full JSON Schema
Draft 2020-12 compliance. Without the feature, the existing lightweight
builtin validator is used.

Closes #351
